Rename EnittyStore to EntityCache (EntityStore already existed)
Also makes EntityCache an interface, to seperate the interface from
the implementation.
